+++ changelog.md
@@ -3,6 +3,7 @@ See also http://pvp.haskell.org/faq
 0.6.0.0 *TODO*
 
   * Add offending path as new field to `TarBombError` constructor
+  * Add `Traversable Entries` instance
 
 0.5.1.1 Herbert Valerio Riedel <hvr@gnu.org> August 2019